lab_1/clients/requirements.txt
astroid==3.3.5
azure-core==1.32.0
azure-cosmos==4.7.0
azure-data-tables==12.5.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
idna==3.10
isodate==0.7.2
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pylint==3.3.1
requests==2.32.3
six==1.16.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
